字段查询优化

+++ beetlsql-service/src/main/java/com/rust/demo/entity/Enemy.java
@@ -17,7 +17,7 @@ public class Enemy {
 
     private String enemyTags;
 
-    private Byte sortId;
+    private Integer sortId;
 
     private String name;
